artemis/tasks.py: move db initialization out of task_core
According to SQLalchemy docs, the create_engine should be done ideally
only once per process. Having it in task_core causes that we do this
a lot more ...

Move it to module instead, so it is initialized for each worker only
once.

Tested with 60 citool artemis module runs.

Note that there seems to be also other issues, but this seems to resolve
the most pressing one and should bring us more stability to staging
environment for now.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>