db: actually close session
After deploying our latest code to staging, running integration tests and bringing
down Artemis again, I looked again on the code and read the
documentation of SQLAlchemy. If found out that we do not close the
sesion and without it the connection is never freed and each new session
opens a new db connection, so we definitely need to close the session :(

After trying to close the session, I found actually two bugs where the
code is touching the db again after closing it, so this fixes it also.

More reading:
https://docs.sqlalchemy.org/en/13/orm/session_api.html#sqlalchemy.orm.session.Session.close
https://docs.sqlalchemy.org/en/13/orm/session_basics.html#when-do-i-construct-a-session-when-do-i-commit-it-and-when-do-i-close-it
https://stackoverflow.com/questions/21738944/how-to-close-a-sqlalchemy-session

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>