tasks: release guest also in case of errors
It can happen somebody cancels the request earlier it was provisioned,
in this case the request will hand in releasing state.

The same problem can happen if somebody removes guest outside of Artemis
in theory.

I feel it would be better to remove the guest also in these cases,
artemis tries to remove, if it fails, it should not be a fatal problem.

If we want to keep released guests, let's keep them in DB ...

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>