container: add artemis logging directly in entrypoint script
Previously we added support for running entrypoint via bash and teeing
the output to a log file directly in openshift template. This turned
out to be a bad idea, due to the fact that it completely broken
the SIGTERM handling of the pod.

Let's move the logging into entrypoint directly and let ARTEMIS_LOG_DIR
specify the directory to log into instead. This should fix the currently
seen problem in SIGTERM handling and get Artemis redeployments back into
a normal state.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>