retries: move from 'routing' to 'error' if retries are over.
It can happen we were not able to provision the machines in any pool
and no suitable pools were found, thus we stay in routing state.
In reality we want to move the guest into 'error' state in this
occasion.

This should put (partially) an end to instances staying in routing
state.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>