acquire_quest: do not cancel succesfully provisioned machine
Seems this cancel handling is wrong, we successfully waited for
the machine to be provisioned, then we undo that action if cancel
was set, return the machine and take no action. This means the request
will stay in provisioning and it is never retried really.

I believe we can safely remove this and finish up succesfully the
provisioning. Found during our stress testing.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>