drivers/aws: add support for provisioning without spot request
This merge requests makes it possible for the driver to use normal
provisioning, without asking for a spot request. This is useful
as another pool for fallback from spot requests, which are a limited
resource.

To use provisioning without spot requests, do not specify the option
`use-spot-request` in the pool configuration.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>