Fetching events moved into db.GuestEvent
Fetching DB records seems to be a job for the record model class, API
manager should use its internal API. Fetching events is obviously called
not just from API, but from the hook as well, therefore the manager does
not seem to be the right place for this.

Primary trigger for this patch to split GuestEventManager and fetching
events is the need to avoid using the manager from hooks - hooks shall
be given the session from their caller, never should they acquire their
own session, which is exactly what manager does because it has no parent
caller. Hooks, on the other hand, are always executed in the context of
already existing session (and other objects, like logger).