metrics: refactored to avoid global vars and names compensating for namespaces
I was increasingly unhappy with the direction our implementation of
metrics were heading:

* global variables - all Prometheus objects exist as global variables,
and we need locking to avoid overwriting values by simultaneous API
requests.
* flat namespace - all Prometheus metrics for different parts of our
system lie next to each other, and the namespaces are hardcoded in their
names.
* if I'd lke to inspect a metric, e.g. in a routing script, I'd have to
execute an ad-hoc query to fetch it, which requires me - as a
administrator of an Artemis deployment - to know Artemis internals on a
very low level.

This patch tries to solve ^^ with this:

* all metrics are now grouped by their system in several data classes.
There are no naked metrics, no global variables.
* these classes provide methods to load their data from the database -
if I'd want to use metrics in my hook script, I'd call this method
without worrying about what's the metric name in the database.
* these classes are used to pass metrics in our code, exclusively.
* dependency injection: all methods dealing with metrics expect caller
to provide context like logger, session and so on, instead of guessing
or using global objects.
* each class is responsible for converting its metrics to Prometheus
format.
* and, finally, these classes shall provide helper methods for updating
their metrics. Instead of constructing queries everywhere,
`MySystemMetrics.inc_usage()` shall be used to keep our code clean.