README.adoc: improve no minishift dev env guide
I had hard time to understand docker creates volumes which
need to be specially removed to reinitialize the database.

Update the guide to use detached mode for docker compose, which
makes more sense in the guide steps and mention explicitely usage
of `docker-compose down` and also the `-v` option to remove volumes,
among other things.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>