Fixes submission of tracebacks to Sentry
Gluetool is using now-deprecated library, Raven, to talk to Sentry. It
has some limitations, and our approach to logging tracebacks did not
really work.

If there is a traceback, it must be converted to a JSON understood by
Sentry - Raven could do it automatically for us, but it cannot cope with
Python 3.5+ FrameSummary objects.