metrics: implement succesful failover counting
We need to be able to count also successful failovers. This MR
implements the support for this.

After a succesful provision we check for failed pools before and raise
their counter. If there were no failed pools, we do nothing.

To be able to get previously failed pools, we need to get to events for
the given guest, so add to workspace a helper function to do exactly
that.

Resolves #69

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>