aws: add possibility to adjust root disk size
By default AWS creates disk size according to the AMI.
This is usually quite small, depending on the image, usually
around 6GB. To resonably run our workloads we need
to be able to use larger disks.

This merge request adds and new optional `default-root-disk-size`
configuration option to the aws pool config, which can adjust
the root disk size. To do that we first need
to find the default block device mapping of the image, adjust
it and pass on when creating the spot request.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>