Drops Guest/Snapshot classes in favor of direct use of their DB representation
`Guest` and `Snapshot` classes were supposed to be higher-level
extensions of DB models, to be passed between Artemis core and drivers.
But, all we need to store are pool-specific data, like OpenStack
instance ID, which we already store in `Guest.pool_data`: to achieve
that, we don't need full-blown class.

Pool drivers are now given `GuestRequest` and `SnapshotRequest`
instances. This removes on elayer of abstraction, and makes it easier
for us to reason about objects used and their relations.

In future patches, this will become quite handy when we clear rules of
when and how to detect reality of pool backend does not match Artemis'
knowledge, and dealing with irrecoverable failures raising from this.

It will also become much easier to implement lock-free tasks by means
of "generation" mark on these two objects - the additional level makes
reasoning about generations needlesly complex.