Decouples pool resource release from guest release
Since the beginning, "release guest" meant also "release all resources
allocated for the guest". This is not necessarily always the best way to
go, because it blocks workflow: guest update may need to drop existing
instance and acquire a new one, and the instance release may fail and
actually block this sequence.

If we decouple these two actions, we could "drop instance" in our update
chain, which would mean only scheduling new task for actual release of
resources, and the update chain would be free to proceed. It would also
open new possibilities in other places where we might need to perform a
cleanup, before we ask routing to pick new pool. This would not be
possible, as "release guest" could block on broken pool, effectively
preventing us to return to routing. Having this as a fire-and-forget
task unties our hands.