Adds prototype of tests for DB functions
To bootstrap tests for functions that work with DB, I added a "safely
update DB record" primitive - there is similar function,
`safe_db_execute`, but that one does not match its name, and is suited
for removal once replaced with `safe_db_update`.

To support the tests, we need fixtures to construct DB, initialize it,
skip tests based on the actual DB implementation, and so on.