Converts "revert to routing" code to task-like actor
It is already using tools our tasks are using, and it needs more. So,
instead of importing bits and pieces to middleware, the code now lives
in tasks, and it is using Workspace and handling of failures and other
stuff, making it behave like a task.

The problem is this: "return to routing" is open to leaks. It is
possible to run out of retries on a provisioning task, and middleware
would than return guest to routing - without releasing already acquired
resources.

It is less likely with some pools: OpenStack and AWS request an
instance, and then wait for it, it's rather unlikely they would even
successfully request the instance should the pool was unstable. But,
some drivers can easily succeed submitting their requests, like Beaker:
with much longer waiting time, Beaker may easily submit a job, and than
the pool can break down, making `update_guest` tasks to crash and run
out of attempts.

So we do have to take care of releasing anything we might have acquired
- some drivers won't have anything to do, but some need this
opportunity. Therefore the code is now refactored and asks pools to
release guest data before falling back to routing.

Now this means the code should use `Workspace`, test for issues,
failures, realize it can fail. With the recent decoupling of resource
release into a separate task, we can now easily release guest resources
without actually talking to the pool in this path, making it much safer
to cleanup when we want to switch pools.