Makes handling of retries more open to special handling of tasks not involved in provisioning
Some of these tasks, like `release_pool_resources`, may still accept a guestname as its parameter, to keep track of guest request actions, but they are not part of the provisioning, therefore "revert to routing" would be pointless.

The patch reorders code a bit, so we can avoid running provisioning tail handler for these tasks, and add special handling for those that need it in the future.