Adds --against-db-url option for tests, to support more dialects
To allow testing of DB-touching code with more than just SQLite - PostgreSQL is our DB of choice, btu we do not test our code with it... - the patch adds an option which allows parametrization of the `db` fixture. This fixture covers connecting to a database, and it is now possible to add more databases to its parameters. We keep SQLite as a default, but now we can do `--against-db-url=posgresql://user:password@test-postgres-deployment:port` and get our code tested against an ad hoc PostgreSQL.