Adds support create/drop database in tests
Based on the URL, if the dialect isn't SQLite, test harness tries to
crate and drop the database used for tests. It does not try to setup the
database server, or grant privileges to the user - that's up the the
developer running the test.

This aims at running tests against PostgreSQL, which will need a bit
more setup before running the test suite, and it needs to reset the
schema before each test.