Adds SafeQuery, Result-returning successor of Query
Query has a problem: when its methods are executed, and touch the actual
db, they may run into troubles and raise exceptions. They would not be
unhandled, but probably a more distant layer would take care of them,
loosing access to the context of the queries.

Instead, we need a "safe" query class, a class that plays well with the
rest of our code base, returning Result instances and catching
exceptions. This patch adds it, and also changes one use of Query into
SafeQuery one (KnobSourceDB).