Adds opt-in linters for docstrings
Based on flake8-docstrings (pydocstyle wrapper) and darglint. There is some overlap of coverage, but both tools lack something the other can do: darglint wouldn't warn on missing docstrings, and flake8-docstrings does not lint the actual content of the docstring, i.e. whether all parameters and return values are documented, and so on. Both work with flake8, therefore it is trivial to use them at the same time.