Resets flake8 ignore list to "report everything"
The default list does have few items, but none is so important that it
should remain hidden.

While working on docstring linters, it appeared the ignore list must be
tweaked, and that would be much easier if we start with a clean slate.

This reset revealed few minor issues in our code - indentation mostly -
which should be fixed, actually, all of them. So it seems an empty
ignore list bring no harm.