Changes hooks to accept keyword arguments only
We used keyword arguments to make parameters more visible when passing
them to hooks, but that sort of introduced the need for a default value
and the optionality of non-`None` value. With our current hooks, the
values can never be `None`, therefore the `Optional` and `assert` are
just causing a mess.

Switching to keyword-only approach let's us drop the default values
while keeping names clearly visible, and doesn't prevent us from adding
more parameters or default values where needed.