Refactors pool capabilities to introduce supported architectures
Pools share some basic capabilities from which we picked snapshots, but
we also need supported architectures to avoid hardcoding arches in
drivers. Since these capabilities are common to all drivers, it is
therefore better to handle them in driver base class as much as
possible. This does not prevent drivers to override values deduced by
the base class.

This involves a bit of refactoring to unify capabilities in the
configuration as well, to give the base class an opportunity to handle
common capabilities without involving child classes.