Adds environment among well-known Failure arguments
It is often attached to failures, and gives us chance to enforce type.
Which is probably not crucial since our `Environment` class is actually
a dataclass which can be easily serialized into JSON. When printing or
storing in guest events, it is serialized to JSON and string
automagically. But it will matter when one tries to reason about methods
of `failure.details['environment']` - from now on it will be always one
type and nothing else.