Refactors how Artemis handles image name and ID
Instead of pure string, the patch introduces a dataclass to capture both
image name and corresponding ID. The string is indeed handy, but it's
not possible to prevent the case of mistaken identity: return name where
ID was expected. This makes for lousy interfaces: if it's not important
to distinct between name and ID, we should drop one of them, and if it
is important, we must make sure they are never exchanged without
noticing.

Type annotations can help, and the structure adds a space for
documentation as well.

The patch puts the structure to use right away, by introducing
`PoolDriver.image_info_by_name` that returns the structure for a given
image name. This is effectively what our pool hooks do, and it is task
generic enough that should land in the drivers itself, instead of being
implemented in hooks - this prevents the reuse of the "name => id"
mechanism, and we rellay want to access this feature in future patches.