Converts "compose => image name" mapping failures as irrecoverable
As discussed in #71, failure to map compose to image name should be
considered as an irrecoverable error. Unlike the failure to find image
in the pool - this error remains recoverable.

This should help us implemnent proper `PoolDriver.can_acquire` method in
existing drivers, and shorten error paths when user requests unsupported
compose - as oposed to requesting supported compose, but failing to find
it in the pool: this situation is deemed recoverable with the help of
administrators.

Implements #74.