Fixes handling of safe_db_change in API methods
Handling of changed/not changed result of `safe_db_change` call was
fragmented: 3 call sites did the same thing in a slightly different
manner, and 2 of them were even (mistakenly) ignoring possible outcome.

For some queries, it is impossible to tell the actual cause they did
not apply. It could be a conflict or missing record. Picking `409
Conflict` to report these, because this signals user they should
re-examine the situation (making sure the guest exists and that there
are indeed no snapshots) and try again (removing the guest).