Refactors metric names and descriptions
The naming of some metrics was quite unusual, slightly inconsistent with
the rest of the pack. This patch tries to polish these bits, bringing a
bit of order into names and descriptions.

And to make it more readable, adding a short page documenting these
metrics and where to find them, so we can keep the law and order.