Adds `ctime` property of guest requests
Set to UTC "now" when the guest request is created. The same information
is available in guest events, but one has to perform joins and filter
events - it seems easier to denormalize a bit here, for the same of
simpler and cheaper queries.

One of the goals is to trackage of guest requests, to discover stalled
ones.