Adds context variables for the most common objects we pass around
To manifest the use, patch converts `Metrics.sync()` method to use the
context variables instead of being given logger, DB and session as
arguments.

See [1] [2] for details, but this approach seems to be very fitting to
how our code works: everything starts in the limited set of entry
points, these create loggers, sessions and other objects representing a
"virtual context" which are then passed to the rest of the code.

Yes, it looks like global variables, but only the names are shared, the
values are not. I was originaly thinking about a class bundling all
these objects together, to pass just this instead of all the names, but
this seems to be also a viable option, and it's a standard library.

* [1] https://docs.python.org/3.7/library/contextvars.html
* [2] https://www.python.org/dev/peps/pep-0567/