Adds `with_context` decorator for simplified injection of context variables
Another step of the context development: instead of accessing context
variables directly, we can use a decorator for a simplified dependency
injection. It is limited in what callables can be decorated - only those
accepting positional arguments - because of my inability to express
more complex cases with type annotation, but even this should be enough
to help *a lot* with metrics and tasks.