Adds a task for periodical refresh of pool image info
Data are cached and refreshed periodically by task running for each
pool.

Why: we discovered that fetching the list can take quite some time,
dozens of seconds. It is repeated often, each guest request provisioning
attempt runs it. It (relatively) rarely changes. Especially when dealing
with multiple requests at the same time, having a cached image info can
shorten the provisioning by a minute *and* be more friendly to the whole
infrastructure - no burst of image list API requests.

The use of cache is not mandatory: drivers may or may not use it, it's
up to their authors. Some drivers might even ignore it completely
because they don't need this image info concept. But when needed, the
cache is here and it's trivial to join.