Adds `ArtemisGuestName` tag unifying the shattered guest naming
Azure and OpenStack drivers were naming their guests, using the same
schema. AWS uses different schema, with an IP address. This patch adds
a guest-level tag providing this name for drivers to use. Azure and
OpenStack are now updated to do so, AWS is left untouched because I
need to investigate the importance of the IP address.