Allows passing any arguments to functions decorated with `with_context`
Under the hood, the decorator does exactly tha anyway: takes incomming
arguments, adds dependencies and calls the decorated function. The
problem is with the signature: according to my knowledge, it's not
possible to properly express such a decorator, one that changes
signature of the decorated function by removing some of the parameters.
Especially parameters that are purely optional - not just their values,
but the parameters themselves.

The standing limitation allowed only one argument to be passed to the
decorated function, representing `self` object, but this was artificial
and did not really help with the typing problem. Therefore dropping this
limitation to allow any callable to be decorated, and we will have to be
cautious and wait for necessary features to land in typing, mypy and
other packages.