Releases DB session in refresh-pool-image-info as early as possible
This task 1. can take long time as it talks to the actual cloud, and
getting a list of images may take even tens of seconds, 2. needs DB
session to get the pool instance only and that's it.

After the pool instance is instantiated, the session is no longer
needed - but will remain opened, increasing number of "idle in
transaction" connections. Tasks do not have any infrastructure for
preliminary session release, yet, therefore adding an explicit rollback
here, which will be polished (together with context stuff already used
by the task) - the same action would be handy in few other tasks as
well.