Adds ``context`` context manager to set context variables within a block
In the example bellow, ``foo`` will be called with
``some_cache_instance``:

```python
@with_context
def foo(cache: redis.Redis):
  ...

with context(cache=some_cache_instance):
  foo()
```

``context`` will take care of unrolling the change when leaving the
context.

The idea is to use this manager to "spawn" a context e.g. in actors and
doers, instead of managing each context variable alone which would mean
too much work to undo the context spawn.

Not yet used, but it will be once we dive into tasks.