Adds limit and usage tracking for OpenStack networks
OpenStack driver tries to pick the most unused network, modulo a network
name pattern. The list of networks is fetched for every provisioning
attempt. THis patch moves this operation to the metrics refresh
workflow, which 1. allows us to track the network usage in the same way
as other pool resources, and 2. allows driver to use this cached
information when picking the network, instead of fetching it especially
for each request.

This makes the network usage available to routing, too, because it's
transparently handled by pool resource metrics and its "which resources
are depleted?" methods.