Moves context code into a separate module
This is part of solution of a nastry circular dependency. `__init__.py`
imports `db.py` - mostly for typing, but also to include `db.DB` as a
type in `CONTEXT_PROVIDERS` - and `db.py` imports from `__init__.py`,
making it hard to avoid circular imports.

Usualy it's easy to import just before one needs the given object, but:
I'd like to drop vault usage in favor of storing the data in DB directly,
encrypted. SQLAlchemy is capable encrypt/decrypt transparently, but it
requires the password *when the model class is declared* - and here's
the problem, it's not possible to import `db.py` from `__init__.py` and
at the same time import `KNOB_VAULT_PASSWORD_FILE` from `__init__.py`
while we're still just loading code.

Therefore, to break the cycle, moving context code into a separate
module, which removes the only use of a real `db.py` object in
`__init__.py` - the rest is just the typing stuff, and we can live with
type annotations being strings.