Adds task to run post-install script when driver cannot handle it
Driver can now announce whether it's capable of handling post-install
script, e.g. by using cloud-init. If the driver cannot run the script,
Artemis now gains a task to upload and execute the script.

This should bring Beaker driver up to the same level as the rest of the
drivers by closing this implementation gap.