Adds helpers for work with remote content of guest filesystem
We already used SSH to connect to the guest, this patch converts this
bit into a standalone function, and adds upload/download helpers using
the same approach.

The motivation: I'd like to extend support of post-install script to
drivers that don't have native support, this requires uploading the
script and running it. Here's where we'd use the helpers.