Fixes possible optionality of parsed broker port
Revealed by mypy: `parsed_url.port` can be `None`, if left unspecified. To make mypy happy, we need a value to use in such a situation.

On purpose, this default port is not handled by knob. We already have a knob for broker URL, this default value is merely an implementation detail we need to resolve user not providing a port.