Adds support for custom in-pool instance naming
Artemis uses a default naming scheme to name newly crated in-pool
instances that represent each guest. This patch allows user to specify
their own name when requesting a guest.

Since all pools use arbitrary IDs to identify the in-pool instances, and
these IDs are assigned by the actual pool backend, it shouldn't matter
whether the custom names are unique or not.