Adds support for irrecoverable failures that don't affect the guest request
There are such failures, e.g. in the release-pool-resources chain. If
the release fails, for example because of pool outage, it is not desired
to mark the request as failed by switching its state to `ERROR`. Release
of resources is decoupled, and shouldn't affect the guest anymore - the
guest might have been rerouted to completely different pool by now.