Splits config validation into generic bits and configuration part
In the process, introduces use of `Result` approach instead of
exceptions.

The ultimate goal is to use the generic bits - `validate_data` - in the
process of validation of environment specification, especially its HW
requirements. We do have tools needed, it's just been tied to server
configuration to be usable from other parts of the code.