Moves release guest task dispatch from dispatcher to API server
In the ongoing quest to get rid of the dispatcher, another part of its
work is now moved into API server and hooked to an API endpoint.
Dispatching release-guest-request tasks from the dispatcher under load
can lead to increasing the load even more, almost exponentially. Just
like the `route-guest-request`, it is possible to dispatch the task
once, when the corresponding event happens, and leave the rest to the
broker.