Adds version-prefixed API endpoints
The goal is to introduce version-prefixed API, `/v0.1.2/guests/` and so
on. This is necessary to support API changes without disturbing existing
clients - once in motion, we can introduce changes without affecting
clients that would tie their implementation to a particular API version.

With this patch, Artemis would provide following endpoints:

* `/current/...` - always available, the most up-to-date implementation
* `/vX.Y.Z/...` - with every change, we'd branch original vX.Y.Z API,
  and add changes to `/current`. There will always be a prefix
  corresponding to current Artemis version as reported by
  `tft.artemis.__VERSION__`, which would provide the same implementation
  as `/current`.
* `/...` - current approach where endpoints are living in the top level
  of the hierarchy, with no prefix. Once we are confident all clients
  switch to the versioned scheme, either by accessing `/current` or a
  particular version, we would drop these endpoints.

Implements #37.