Adds a knob for a Vault password
Instead of reading it from a file, support - also - a password provided
directly via environment variable, or, when deployed, from a secret.

The final goal is to drop the file source entirely, and use a proper
secret to provide the password, materialized in containers as an
environment variable. The password must be known as soon as possible, to
allow construction of DB models with encrypted fields. Then, we could
drop the Ansible vault and Ansible, too, and use transparent encryption
provided by [1].

The problem of password file is the fact one has to open the file before
getting to the passowrd - this doesn't play well with DB model
declaration when Python is still loading packages and modules. Hence
constant value, acquired from the environment when the knob is
instantiated.

[1] https://sqlalchemy-utils.readthedocs.io/en/latest/data_types.html#module-sqlalchemy_utils.types.encrypted.encrypted_type