Adds new environment field key dedicated to HW requirements
This is a change of the internal format of the environment, it does
*not* change the public API specification. The API change will follow,
but it can be made as a stand-alone patch, which should make things
easier for reviewers.

The new key, `hw`, should carry HW-related part of the guest request
environment. Current `arch` key, sitting in the top-level area, is now
moved under this new key.