container/Dockerfile: make krb5.conf editable in Openshift
We are fighting a DNS resolution problem in Openshift which
cases issues with resolving SRV DNS records for unknown reason
for Artemis worker - when it runs `bkr` command.

The issue is weird, because using `dig` in another container
works and resolves also the problem in Artemis for a certain
period of time.

Until we know the root cause, let's make `/etc/krb5.conf` editable
to the Openshift user, so we can force a KDC instead of trying to
resolve it in the deployment.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>