Adds full failure details to Sentry submissions
It should be fine, but to be sure to avoid breaking things, also
wrapping the submission into try/except. This should have been done
already, no matter what we submit, but now it's pretty much mandatory.
In case a failure would try to include a weird detail, the submission
would explode with an exception - not acceptable.

Getting more details, on the other hand, is highly needed.