container: move to Fedora base image
As we found out, Alpine does not play nice with Kerberos due to the
missings support for edns0 in musl C library. The problem is hard
to workaround because the resolved does not do UDP -> TCP failover
and also cannot be force do use TCP via the `use-vc` flag, which
is also not around.

So move away from Alpine to mitigate this problem and use a glibc
based OS we all love and hepefully will just work.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>