Changes AWS driver to construct block device mappings from a given flavor
Current flavors lack this information - that will come later - but the
driver is still using the configuration to provide the default root disk
size. This patch reorders the code for clearer integration.

Once we teach the driver to use both custom and existing flavors, we can
drop the default value.

Relates to #132.