Uses Pint package to handle memory/diskspace of flavors internally
While it is trivial to convert MB/GB values to bytes using simple
multiplication, and using Pint to do this could look like a step back
and adding a complexity (true), Pint package will offer more advantages
in the future. Its `Quantity` class would be the single "currency" of
values with units, and it would be used by flavors, environment and
metrics, allowing easy to use conversions. It is also capable of very
smart string parsing which will come handy when decoding user-specified
environment constraints.

https://pint.readthedocs.io/en/stable/tutorial.html