Drops notion of "moving" API versions, Artemis API will follow more common versioning
So, it's `v0.0.18` since that's what we have *now*, and in the future,
versioning won't try to keep transient versions in play. E.g. we're
having `v0.0.18`, I suspect there will be some changes in v0.0.19, but
then another changes would come in Artemis v0.0.23, for example, and
there will be no `/v0.0.20`, `/v0.0.21` nor `/v0.0.22`.