Adds default implementation of base metrics class methods
Instead of raising an `NotImplementedError` exception, the base class now tries to:

* find fields that are descendants of this base class,
* when `sync()`, `register_with_prometheus()` and `update_prometheus()`
  are called, it propagates this call down to these fields.

This behavior now saves some work to child classes: since they all form
a tree of metrics containers, some nodes in this tree had to explicitly
call these methods for all their children nodes that were also
containers. The default implementation does this automagically.

This patch is a first of a series focusing on moving metrics more to the
declarative side, providing automation for the most common tasks, so we
can add more metrics with ease, and make their definitions more
readable.