Clears API versioning implementation
* the milestone API versions are still with us
* dropping "compatible" versions except of the very special ones,
  "current" and "toplevel"
* handle environment schema versioning - load the one matching the API
  version

This should be the almost last step in clearing this mess - all
redirects in place, milestone versions clearly visible, environment
schema tied to its API version. The next step - possibly the final one -
would be to use classes and inheritance to reduce the amount of code
used to generate routes when API changes.