Adds basic support for flavor modifications via config file
This patch, together with one adding support for custom flavors, should
provide the flavors rich enough to be usable with HW requirements
currently defined by TMT.

Some information is not available via cloud API, therefore the driver
cannot extract it when fetching flavor info. For example, AWS documents
what kind of CPUs are provided by each flavor, but it's not possible to
find this information in the output of `describe-instance-types`
command.