brew-build-task-params: map SERVER parmeter via rules file
We have been using primary_task.ARTIFACT_NAMESPACE in the SERVER
value, what was never the correct thing to do, but it worked.

With change in

https://pkgs.devel.redhat.com/cgit/tests/distribution/commit/?id=a72f323d2b09c92e389659c915ac2d6ba6ff49e4

this logic changed, and all pipelines broke :/ It tooks us while to
realize the underlying problem.

Introduce for the paraemter a rules file, so we are prepared for the
future values of this option, now that CentOS Stream testing is knocking
on the door.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>