Splits handler of "create guest" API request per version
Since this API endpoint is the one that changed over the time - because
the environment specification changed - its handler should be split into
distinct implementations. This allows proper schema to be applied.

This is the expected process - route generators will be more or less the
same, until some of them changes. Then we spawn custom handler, but we
point it to shared, parametrized code.