Adds "format struct as YAML" helpers for logging
We do have `log_dict` provided by Gluetool, but, sometimes, YAML is more
readable for humans. I realized I often depend on checking logs when it
comes to provisioning, images and flavors, and unformatted JSON is just
a pain to read.

Adding a helper to format structure as a YAML, and using it in drivers
that log image and flavor used for provisioning, to provide a nicely
formatted output.

Hopefully, I'll be able to move this functionality to Gluetool, possibly
as a switch of `format_dict` and `log_dict`.