Adds `from_failure` helper for creating chains of failures
More or less a mere syntax sugar, but it adds nice semantic way to
create a high-level Failure, wrapping the low-level one without any loss
of information.

And preserves ``recoverable`` automagically, which is **not** true when
`caused_by=...` is used. This way, irrecoverable failures are propagated
even when wrapped with more context.