Adds proper value casting to knob API
This was a known issue: values passed to `knob set ...` command were
stored as-is - no conversion was applied, everything was a string. This
was a serious issue, preventing any real-life usage - e.g. a knob that's
supposed to return `int` would, after such a change, return a string.

API server was not aware of knobs' properties, including the functions
used for casting envvar value to correct type. To make this information
available, this patch:

* records all knobs that can be changed via API - knobs created with
`has_db=True`. No other knobs can be changed via API, e.g. to change a
knob backed by envvar alone, one needs to restart the process.
* the casting function is no longer specific for envvar source, since
it's used also for knobs backed by DB.
* API can now inspect this collection of knobs, and find the correct
casting function. The function is then a) reported in output, to make
clear what value is expected, and b) used to convert the incoming
string to correct type.

Per-pool knobs are a bit special since their DB records don't point to
the `Knob` class instance in the code - the `knobname` field in DB is no
longer `pool.$poolname.foo` but the placeholder `$poolname` is replaced
with actual pool name. For these records, we must search a bit harder to
find the actual `Knob` instance, by stripping away the `pool.$poolname`
prefix.

During the refactoring, it become possible to report the "static" value
of per-pool and DB backed knobs: if they have other sources, like the
default value, it is possible to deduce the value during bootstrap time.
It can't be used for any real decission, but it represents the default
value of these knobs if there's no DB record - which is the effective
value *now*, until user intervenes via `knob set ...` command.

So, after this patch, API correctly converts incoming values to their
respective types, or raises an error, and reports correct values when
used did not modify the knob yet.