Modifies DB initialization to proceed and mirror config into DB
The DB initialization process was not repeatable: it ran once, and every
other attempt failed because of conflicts. Which is far from ideal. This
patch switches from `session.add` to `upsert()` calls which has two
advantages:

* works without conflicts
* configuration file is now the single source of truth: when config->DB
  transfer works just when running for the first time, configuration is
  more or less ignored, and runtime changes can get easily out of sync.

With this patch, what's in configuration get deployed to DB Artemis
starts. To change runtime settings, Artemis provides knobs which are
*not* modified by the configuration.